feat: extract pure pdf planner
